added align option for features list
